WMV-363 : update wow library for x64

+++ src/games/wow/CASCFile.cpp
@@ -112,7 +112,7 @@ bool CASCFile::isAlreadyOpened()
     return false;
 }
 
-bool CASCFile::getFileSize(unsigned int & s)
+bool CASCFile::getFileSize(size_t & s)
 {
   bool result = false;
   
@@ -149,7 +149,7 @@ void CASCFile::doPostOpenOperation()
     if (std::find(KNOWN_CHUNKS.begin(), KNOWN_CHUNKS.end(), magic) != KNOWN_CHUNKS.end() 
         && chunkHead.size <= size)
     {
-      unsigned int offset = 0;
+      size_t offset = 0;
 
       //LOG_INFO << "Parsing chunks for file" << filepath << "First chunk read :" << magic.c_str();
       while (offset < size)
